configured warnings into errors on darwin and fixed several leopard/64-bit wanrings which may have been source of trouble.
git-svn-id: https://subversion.assembla.com/svn/portaudio/portaudio/branches/v19-devel@1304 0f58301d-fd10-0410-b4af-bbb618454e57

+++ src/hostapi/coreaudio/pa_mac_core_blocking.c
@@ -464,7 +464,7 @@ PaError ReadStream( PaStream* stream,
 
     /* report underflow only once: */
     if( ret ) {
-       OSAtomicAnd32( ~paInputOverflow, &blio->statusFlags );
+       OSAtomicAnd32( (uint32_t)(~paInputOverflow), &blio->statusFlags );
        ret = paInputOverflowed;
     }
 
@@ -541,7 +541,7 @@ PaError WriteStream( PaStream* stream,
 
     /* report underflow only once: */
     if( ret ) {
-      OSAtomicAnd32( ~paOutputUnderflow, &blio->statusFlags );
+      OSAtomicAnd32( (uint32_t)(~paOutputUnderflow), &blio->statusFlags );
       ret = paOutputUnderflowed;
     }
